Added 42 triggers
Forced update on the branch for the pull request.

1. Moved the triggers to controllers as @philsturgeon suggested.
2. I tested to check that I didn't mess up things in the admin.
3. But it should be double checked by you guys.

This time I tried following these rules when placing triggers:

1. Use simple and meaningful names
2. Place them where success has been checked
3. Repeat a trigger in places where it is relevant. (like modules)
4. For a create operation, don't pass input data.
5. For a update operation, pass ids.
6. For a delete operation, also pass ids. Might be useful for deleting
references in other tables.

The list of new triggers:

blog_article_deleted
blog_category_created
blog_category_updated
blog_category_deleted

comment_updated
comment_deleted
comment_unapproved

file_uploaded
file_updated
file_deleted
folder_created
folder_updated
folder_deleted

group_created
group_updated
group_deleted

keyword_created
keyword_updated
keyword_deleted

module_enabled
module_disabled

navigation_group_created
navigation_group_deleted

permissions_updated

settings_updated

email_template_created
email_template_updated
email_template_deleted

theme_options_updated
theme_set_default
theme_deleted

user_created
user_updated
user_deleted

widget_enabled
widget_disabled
widget_area_created
widget_area_updated
widget_area_deleted
widget_instance_created
widget_instance_updated
widget_instance_deleted

+++ system/cms/modules/navigation/controllers/admin_groups.php
@@ -78,6 +78,8 @@ class Admin_groups extends Admin_Controller {
 			if ($this->navigation_m->insert_group($_POST) > 0)
 			{
 				$this->session->set_flashdata('success', $this->lang->line('nav_group_add_success'));
+				// Fire an event. A new navigation group has been created.
+				Events::trigger('navigation_group_created');
 			}
 			else
 			{
@@ -109,11 +111,16 @@ class Admin_groups extends Admin_Controller {
 	 */
 	public function delete($id = 0)
 	{
+		$deleted_ids = FALSE;
+		
 		// Delete one
-		if($id)
+		if ($id)
 		{
-			$this->navigation_m->delete_group($id);
-			$this->navigation_m->delete_link(array('navigation_group_id'=>$id));
+			if ($this->navigation_m->delete_group($id))
+			{
+				$deleted_ids[] = $id;
+				$this->navigation_m->delete_link(array('navigation_group_id'=>$id));
+			};
 		}
 
 		// Delete multiple
@@ -121,11 +128,20 @@ class Admin_groups extends Admin_Controller {
 		{
 			foreach (array_keys($this->input->post('delete')) as $id)
 			{
-				$this->navigation_m->delete_group($id);
-				$this->navigation_m->delete_link(array('navigation_group_id'=>$id));
+				if ($this->navigation_m->delete_group($id))
+				{
+					$deleted_ids[] = $id;
+					$this->navigation_m->delete_link(array('navigation_group_id'=>$id));
+				}
 			}
 		}
-
+		
+		// Fire an event. One or more navigation groups have been deleted.
+		if ( ! empty($deleted_ids))
+		{
+			Events::trigger('navigation_group_deleted', $deleted_ids);
+		}
+				
 		// Set the message and redirect
 		$this->session->set_flashdata('success', $this->lang->line('nav_group_mass_delete_success'));
 		redirect('admin/navigation/index');
